+++ car-purcharse-service/src/main/java/com/hcltech/car_purcharse_service/dto/CarDto.java
@@ -4,70 +4,70 @@ public class CarDto {
 
     Integer id;
 
-        String make;
+    String make;
 
-        String model;
+    String model;
 
-        int year;
+    int year;
 
-        Double price;
+    Double price;
 
-        Boolean isAvailable;
+    Boolean isAvailable;
 
-        Long sellerId;
+    Integer sellerId;
 
-        public CarDto() {
-        }
+    public CarDto() {
+    }
 
-        public CarDto(Integer id, String make, String model, int year, Double price, Boolean isAvailable, Long sellerId) {
-            this.id = id;
-            this.make = make;
-            this.model = model;
-            this.year = year;
-            this.price = price;
-            this.isAvailable = isAvailable;
-            this.sellerId = sellerId;
-        }
+    public CarDto(Integer id, String make, String model, int year, Double price, Boolean isAvailable, Integer sellerId) {
+        this.id = id;
+        this.make = make;
+        this.model = model;
+        this.year = year;
+        this.price = price;
+        this.isAvailable = isAvailable;
+        this.sellerId = sellerId;
+    }
 
-        public Integer getId() {
-            return id;
-        }
+    public Integer getId() {
+        return id;
+    }
 
-        public void setId(Integer id) {
-            this.id = id;
-        }
+    public void setId(Integer id) {
+        this.id = id;
+    }
 
-        public String getMake() {
-            return make;
-        }
+    public String getMake() {
+        return make;
+    }
 
-        public void setMake(String make) {
-            this.make = make;
-        }
+    public void setMake(String make) {
+        this.make = make;
+    }
 
-        public String getModel() {
-            return model;
-        }
+    public String getModel() {
+        return model;
+    }
 
-        public void setModel(String model) {
-            this.model = model;
-        }
+    public void setModel(String model) {
+        this.model = model;
+    }
 
-        public int getYear() {
-            return year;
-        }
+    public int getYear() {
+        return year;
+    }
 
-        public void setYear(int year) {
-            this.year = year;
-        }
+    public void setYear(int year) {
+        this.year = year;
+    }
 
-        public Double getPrice() {
-            return price;
-        }
+    public Double getPrice() {
+        return price;
+    }
 
-        public void setPrice(Double price) {
-            this.price = price;
-        }
+    public void setPrice(Double price) {
+        this.price = price;
+    }
 
     public Boolean getAvailable() {
         return isAvailable;
@@ -77,11 +77,11 @@ public class CarDto {
         isAvailable = available;
     }
 
-    public Long getSellerId() {
+    public Integer getSellerId() {
         return sellerId;
     }
 
-    public void setSellerId(Long sellerId) {
+    public void setSellerId(Integer sellerId) {
         this.sellerId = sellerId;
     }
 }
